Split lax::alloc submodule
